Implement ICE40 PLL dynamic reconfiguration
"Test mode" allows a word to be shifted in to reconfigure.

Adds config registers exposing PLL resets, lock status, and shift reg controls.
Add a proper reset to the pclk domain.

This also implements a simple (Makefile-driven) fixed clock multiple, which can
be used or bypassed.  This is plan B in case the PLL DR approach is bad.