Increase output X counters (and line buffer size) to accommodate 1056x256 8bpp mode
Quite costly for one mode, in that the line buffer goes from 2KB to 4KB!
This could be shrunk to 3KB with additional fiddly logic to deal with buffer chunks,
in future work.

However, it's useful to deal with these "wide" modes because there may be some interesting
future 16BPP modes such as 640x256 which will need the extra buffering.